对请求进行了优化，组件进行了进一步抽离，简化优化了部分代码
